+++ src/NotepadNextApplication.cpp
@@ -29,6 +29,7 @@
 #include "ApplicationSettings.h"
 #include "NppPluginManager.h"
 #include "LspManager.h"
+#include "SchemaManager.h"
 
 #include "LuaState.h"
 #include "lua.hpp"
@@ -132,6 +133,7 @@ bool NotepadNextApplication::init()
     recentFilesListManager = new RecentFilesListManager(this);
     editorManager = new EditorManager(settings, this);
     lspManager = new LspManager(editorManager, this);
+    schemaManager = new SchemaManager(editorManager, this);
     sessionManager = new SessionManager(this);
 
     connect(editorManager, &EditorManager::editorCreated, recentFilesListManager, [this](ScintillaNext *editor) {